requirements/requirements.txt
asgiref==3.7.2
awscli==1.33.5
boto==2.49.0
boto3==1.34.122
botocore==1.34.123
Brotli==1.1.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cssselect2==0.7.0
Django==5.0
django-ckeditor==6.7.0
django-environ==0.11.2
django-js-asset==2.2.0
django-model-utils==4.4.0
django-storages==1.14.3
djangorestframework==3.15.1
docutils==0.15.2
EasyProcess==1.1
fonttools==4.53.0
hashids==1.3.1
html5lib==1.1
idna==3.7
imgkit==1.2.3
install==1.3.5
jmespath==0.10.0
pillow==10.2.0
pkg_resources==0.0.0
psycopg==3.1.8
pyasn1==0.6.0
pycparser==2.22
pydyf==0.10.0
pyphen==0.15.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
PyVirtualDisplay==2.0
PyYAML==6.0.1
qrcode==7.4.2
requests==2.32.3
rsa==4.7.2
s3transfer==0.10.1
setuptools==44.0.0
singlemodeladmin==0.9
six==1.16.0
sqlparse==0.4.4
tinycss2==1.3.0
typing_extensions==4.10.0
urllib3==2.2.1
weasyprint==53.2
webencodings==0.5.1
wheel==0.34.2
zopfli==0.2.3
